RenderProxySliver (#45942)
