+++ packages/flutter/lib/rendering.dart
@@ -50,6 +50,7 @@ export 'src/rendering/paragraph.dart';
 export 'src/rendering/performance_overlay.dart';
 export 'src/rendering/platform_view.dart';
 export 'src/rendering/proxy_box.dart';
+export 'src/rendering/proxy_sliver.dart';
 export 'src/rendering/rotated_box.dart';
 export 'src/rendering/shifted_box.dart';
 export 'src/rendering/sliver.dart';
